backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.8
qdrant-client==1.6.9
psycopg[binary]==3.1.15
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
pydantic==2.5.2
pytest==7.4.3
pytest-asyncio==0.21.1
